Refactor to merge the command manager service into the Doulevo class, the separate service isn't really needed.
